# Timeseries: support panning