… test/fixtures/scale.timeseries/source-data-offset-min-max.png
… test/fixtures/scale.timeseries/source-labels-offset-min-max.png
… test/fixtures/scale.timeseries/ticks-reverse-max.png
… test/fixtures/scale.timeseries/ticks-reverse-min-max.png
… test/fixtures/scale.timeseries/ticks-reverse-min.png