Resolves #9344

This changes how the values outside the data range are positioned. Previously they would in proportion to the nearest step, including the scale min/max. Now they are positioned relative to the whole visible data-span.

Summary:
* values outside the actual data points are not included in the interpolation table. This includes min/max when they extend the range.
* interpolating the outliers is done relative to the whole (visible) data range, to keep the scale consistent when panning.
* The "normalized" optimization is removed, because the data index is not relative to the scale range when min/max are outside data. This optimization would have also failed when there are timestamps from both labels and data or when the scale is zoomed, because data indices would not match the lookup table indices.