@kurkle 2021-07-03
I did some benchmarking (chrome, uplot bench, 100 runs): 

master:
| normalized | master | pr | updated pr |
| --- | --- | --- | --- |
| false | 111ms | 111ms | **73ms** |
| true | **52ms** | 86ms | 54ms |