@kurkle 2021-07-03
| Filename | Size | Change |
| :--- | :---: | :---: |
| `dist/chart.esm.js` | 71.2 kB | +122 B (0%) |
| `dist/chart.js` | 90.3 kB | +123 B (0%) |
| `dist/chart.min.js` | 64.4 kB | +67 B (0%) |